Feature(invite): "Show invite dialog"
- When creating a class, the invitation link dialog is displayed.

related to : #102
