+++ app/src/main/java/com/dnd/sixth/lmsservice/di/modules/UseCaseModules.kt
@@ -1,10 +1,13 @@
 package com.dnd.sixth.lmsservice.di.modules
 
 
+import com.dnd.sixth.lmsservice.domain.useCase.GetDailyClassTimeLineUseCase
 import com.dnd.sixth.lmsservice.domain.useCase.dailyclass.CreateDailyClassUseCase
 import com.dnd.sixth.lmsservice.domain.useCase.dailyclass.DeleteDailyClassUseCase
 import com.dnd.sixth.lmsservice.domain.useCase.dailyclass.GetDailyClassListUseCase
 import com.dnd.sixth.lmsservice.domain.useCase.dailyclass.UpdateDailyClassUseCase
+import com.dnd.sixth.lmsservice.domain.useCase.feedbackComment.GetFeedBackUseCaseFromDaily
+import com.dnd.sixth.lmsservice.domain.useCase.feedbackComment.UpdateFeedBackUseCase
 import com.dnd.sixth.lmsservice.domain.useCase.push.comment.ChangeCommentPushUseCase
 import com.dnd.sixth.lmsservice.domain.useCase.push.comment.GetCommentPushUseCase
 import com.dnd.sixth.lmsservice.domain.useCase.push.feedback.ChangeFeedbackPushUseCase
@@ -26,40 +29,12 @@ import com.dnd.sixth.lmsservice.domain.useCase.user.number.parent.SaveLocalParen
 import com.dnd.sixth.lmsservice.domain.useCase.user.number.parent.SaveRemoteParentNumberUseCase
 import com.dnd.sixth.lmsservice.domain.useCase.user.password.ChangePasswordUseCase
 import com.dnd.sixth.lmsservice.domain.useCase.user.username.ChangeUserNameUseCase
-import com.dnd.sixth.lmsservice.domain.useCase.*
-import com.dnd.sixth.lmsservice.domain.useCase.feedbackComment.GetFeedBackUseCaseFromDaily
-
-import com.dnd.sixth.lmsservice.domain.useCase.feedbackComment.UpdateFeedBackUseCase
-
-import com.dnd.sixth.lmsservice.domain.useCase.GetDailyClassUseCase
-import com.dnd.sixth.lmsservice.domain.useCase.feedbackComment.UpdateFeedBackUseCase
-import com.dnd.sixth.lmsservice.domain.usecase.dailyclass.CreateDailyClassUseCase
-import com.dnd.sixth.lmsservice.domain.usecase.dailyclass.DeleteDailyClassUseCase
-import com.dnd.sixth.lmsservice.domain.usecase.dailyclass.GetDailyClassListUseCase
-import com.dnd.sixth.lmsservice.domain.usecase.dailyclass.UpdateDailyClassUseCase
-import com.dnd.sixth.lmsservice.domain.usecase.push.comment.ChangeCommentPushUseCase
-import com.dnd.sixth.lmsservice.domain.usecase.push.comment.GetCommentPushUseCase
-import com.dnd.sixth.lmsservice.domain.usecase.push.feedback.ChangeFeedbackPushUseCase
-import com.dnd.sixth.lmsservice.domain.usecase.push.feedback.GetFeedbackPushUseCase
-import com.dnd.sixth.lmsservice.domain.usecase.push.start.ChangeStartPushUseCase
-import com.dnd.sixth.lmsservice.domain.usecase.push.start.GetStartPushUseCase
-import com.dnd.sixth.lmsservice.domain.usecase.subject.*
+import com.dnd.sixth.lmsservice.domain.usecase.dailyclass.GetDailyClassUseCase
+import com.dnd.sixth.lmsservice.domain.usecase.subject.GetSubjectUseCase
 import com.dnd.sixth.lmsservice.domain.usecase.user.GetUserUseCase
-import com.dnd.sixth.lmsservice.domain.usecase.user.contact.GetLocalContactTimeUseCase
-import com.dnd.sixth.lmsservice.domain.usecase.user.contact.SaveLocalContactTimeUseCase
-import com.dnd.sixth.lmsservice.domain.usecase.user.contact.SaveRemoteContactTimeUseCase
-import com.dnd.sixth.lmsservice.domain.usecase.user.number.myself.GetLocalMyNumberUseCase
-import com.dnd.sixth.lmsservice.domain.usecase.user.number.myself.SaveLocalMyNumberUseCase
-import com.dnd.sixth.lmsservice.domain.usecase.user.number.myself.SaveRemoteMyNumberUseCase
-import com.dnd.sixth.lmsservice.domain.usecase.user.number.parent.GetLocalParentNumberUseCase
-import com.dnd.sixth.lmsservice.domain.usecase.user.number.parent.SaveLocalParentNumberUseCase
-import com.dnd.sixth.lmsservice.domain.usecase.user.number.parent.SaveRemoteParentNumberUseCase
-import com.dnd.sixth.lmsservice.domain.usecase.user.password.ChangePasswordUseCase
 import com.dnd.sixth.lmsservice.domain.usecase.user.profile.GetLocalProfileUriUseCase
 import com.dnd.sixth.lmsservice.domain.usecase.user.profile.SaveLocalProfileUriUseCase
 import com.dnd.sixth.lmsservice.domain.usecase.user.profile.SaveRemoteProfileUriUseCase
-import com.dnd.sixth.lmsservice.domain.usecase.user.username.ChangeUserNameUseCase
-
 import org.koin.dsl.module
 
 val useCaseModules = module {
@@ -85,7 +60,7 @@ val useCaseModules = module {
     타임라인 리스트
      */
 
-    factory{
+    factory {
         GetDailyClassTimeLineUseCase(get())
     }
     factory {
@@ -187,7 +162,7 @@ val useCaseModules = module {
         UpdateFeedBackUseCase(get())
     }
 
-    factory{
+    factory {
         GetFeedBackUseCaseFromDaily(get()) //피드백 내용 불러오기
     }
